+++ api/src/servers/facade.rs
@@ -1,31 +1,12 @@
 // get custom protos
 use proto::roomba_service_protos as protos;
-use protos::roomba_server::{Roomba, RoombaServer};
-use protos::{LightBumper, SensorData, Stasis};
+use protos::roomba_server::Roomba;
+use protos::SensorData;
 use protos::{Odometry, OdometryRequest};
 use protos::{SensorsReceived, SensorsRequest};
-
-// get standard library utils
-use async_std::task;
-use std::marker::Sync;
-use std::pin::Pin;
-use std::sync::{Arc, Mutex};
 extern crate proc_macro;
-use proc_macro::TokenStream;
-
-use futures_core::stream::Stream;
-use futures_util::pin_mut;
-use futures_util::stream::StreamExt;
-
-// gRPC tools
 use crate::servers::endpoints::RoombaService;
-// use futures::{Stream, StreamExt};
-// use proc_macro::TokenStream;
 use crate::servers::utils::SyncBoxStream;
-use colored::Colorize;
-use std::thread;
-use tokio::sync::mpsc;
-use tokio::time::Duration;
 use tonic::{Request, Response, Status};
 
 #[tonic::async_trait]
@@ -43,26 +24,7 @@ impl Roomba for RoombaService {
         &self,
         request: Request<SensorsRequest>,
     ) -> Result<Response<Self::GetSensorDataStream>, Status> {
-        println!("request = {:?}", request);
-
-        let (tx, rx) = mpsc::channel(1);
-        let rx_clone = self.rx.clone();
-
-        let mut count: u32 = 0;
-        tokio::spawn(async move {
-            while let Ok(data) = rx_clone.recv() {
-                thread::sleep(Duration::from_millis(20));
-                println!("{}", "sending data from server".green());
-                tx.send(Ok(data)).await.unwrap();
-                count += 1;
-                println!("count: {}", &count);
-            }
-            println!("{}", "failed sending data from server".red());
-        });
-
-        Ok(Response::new(Box::pin(
-            tokio_stream::wrappers::ReceiverStream::new(rx),
-        )))
+        self.handle_get_sensor_data(request).await
     }
 
     type GetOdometryRawStream = SyncBoxStream<'static, Result<Odometry, Status>>;
@@ -71,6 +33,6 @@ impl Roomba for RoombaService {
         &self,
         request: Request<OdometryRequest>,
     ) -> Result<Response<Self::GetOdometryRawStream>, Status> {
-        unimplemented!("todo")
+        self.handle_get_odometry_raw(request).await
     }
 }
